+++ Assets/Scripts/Entities/FootSoldier.cs
@@ -80,7 +80,7 @@ public class FootSoldier : ContraEntity {
 		}
 	}
 	
-	public override void Damage(int damageTaken = 0) {
+	public override void Damage(float damageTaken = 0) {
 		Destroy (this.gameObject);
 	}
 }
